Add files via upload
This is a test to see if this sends/ works; if it does, this is the greybox of the level. All it needs is a bit of coding for the playerController and the forklifts to move back and forth. Any additional touches you guys want to make can be done in here (barring this works).